feat: Grub reboot improvements
Now grub reboot the system based on the operating system provided by the
user. For now, it only works with hardcoded 'windows' string.

I've tested locally and it worked fine with Ubuntu, KDE neon and Windows
10 on same partition.

The next step is to create a generic function to be called in this
script, resposible to return the ID to reboot.